style: glr.c: prefer returning a value rather than passing pointers
This is very debatable.  This function is not pure at all, so it could
stick to returning void: that's a common coding style to tell the
difference between "real" (pure) functions and side-effecting
subroutines.  However, we already have this style elsewhere (e.g.,
yylex), and I feel the callers are somewhat nice to read this way.

* data/skeletons/glr.c (yygetLRActions): Return the action rather than
passing by pointer.
While at it, fix type of yytoken.
Adjust callers.